feat: 🎸 move common styles into theme
