+++ packages/preset-commonmark/src/node/list-item.ts
@@ -1,5 +1,5 @@
 /* Copyright 2021, Milkdown by Mirone. */
-import { createCmd, createCmdKey, ThemeColor } from '@milkdown/core';
+import { createCmd, createCmdKey } from '@milkdown/core';
 import { liftListItem, sinkListItem, splitListItem, wrappingInputRule } from '@milkdown/prose';
 import { createNode, createShortcut } from '@milkdown/utils';
 
@@ -13,59 +13,40 @@ export const SplitListItem = createCmdKey('SplitListItem');
 export const SinkListItem = createCmdKey('SinkListItem');
 export const LiftListItem = createCmdKey('LiftListItem');
 
-export const listItem = createNode<Keys>((utils) => {
-    const style = utils.getStyle(
-        (themeManager, { css }) =>
-            css`
-                &,
-                & > * {
-                    margin: 0.5rem 0;
-                }
-
-                &,
-                li {
-                    &::marker {
-                        color: ${themeManager.get(ThemeColor, ['primary'])};
-                    }
-                }
-            `,
-    );
-
-    return {
-        id,
-        schema: () => ({
-            group: 'listItem',
-            content: 'paragraph block*',
-            defining: true,
-            parseDOM: [{ tag: 'li' }],
-            toDOM: (node) => ['li', { class: utils.getClassName(node.attrs, 'list-item', style) }, 0],
-            parseMarkdown: {
-                match: ({ type, checked }) => type === 'listItem' && checked === null,
-                runner: (state, node, type) => {
-                    state.openNode(type);
-                    state.next(node.children);
-                    state.closeNode();
-                },
+export const listItem = createNode<Keys>((utils) => ({
+    id,
+    schema: () => ({
+        group: 'listItem',
+        content: 'paragraph block*',
+        defining: true,
+        parseDOM: [{ tag: 'li' }],
+        toDOM: (node) => ['li', { class: utils.getClassName(node.attrs, 'list-item') }, 0],
+        parseMarkdown: {
+            match: ({ type, checked }) => type === 'listItem' && checked === null,
+            runner: (state, node, type) => {
+                state.openNode(type);
+                state.next(node.children);
+                state.closeNode();
             },
-            toMarkdown: {
-                match: (node) => node.type.name === id,
-                runner: (state, node) => {
-                    state.openNode('listItem');
-                    state.next(node.content);
-                    state.closeNode();
-                },
-            },
-        }),
-        inputRules: (nodeType) => [wrappingInputRule(/^\s*([-+*])\s$/, nodeType)],
-        commands: (nodeType) => [
-            createCmd(SplitListItem, () => splitListItem(nodeType)),
-            createCmd(SinkListItem, () => sinkListItem(nodeType)),
-            createCmd(LiftListItem, () => liftListItem(nodeType)),
-        ],
-        shortcuts: {
-            [SupportedKeys.NextListItem]: createShortcut(SplitListItem, 'Enter'),
-            [SupportedKeys.SinkListItem]: createShortcut(SinkListItem, 'Mod-]'),
-            [SupportedKeys.LiftListItem]: createShortcut(LiftListItem, 'Mod-['),
         },
-    };
-});
+        toMarkdown: {
+            match: (node) => node.type.name === id,
+            runner: (state, node) => {
+                state.openNode('listItem');
+                state.next(node.content);
+                state.closeNode();
+            },
+        },
+    }),
+    inputRules: (nodeType) => [wrappingInputRule(/^\s*([-+*])\s$/, nodeType)],
+    commands: (nodeType) => [
+        createCmd(SplitListItem, () => splitListItem(nodeType)),
+        createCmd(SinkListItem, () => sinkListItem(nodeType)),
+        createCmd(LiftListItem, () => liftListItem(nodeType)),
+    ],
+    shortcuts: {
+        [SupportedKeys.NextListItem]: createShortcut(SplitListItem, 'Enter'),
+        [SupportedKeys.SinkListItem]: createShortcut(SinkListItem, 'Mod-]'),
+        [SupportedKeys.LiftListItem]: createShortcut(LiftListItem, 'Mod-['),
+    },
+}));
